+++ src/model/App.tsx
@@ -1,21 +1,18 @@
-/* eslint-disable prettier/prettier */
 /* eslint-disable no-console */
 // eslint-disable-next-line max-classes-per-file
-import { makeAutoObservable } from 'mobx';
-import { SerialPort, SerialPortMock } from 'serialport';
+import { makeAutoObservable, runInAction } from 'mobx';
 
-import NewLineParser from 'util/NewLineParser/NewLineParser';
-import AnsiECParser from 'util/AnsiECParser/AnsiECParser';
 import StopIcon from '@mui/icons-material/Stop';
 import PlayArrowIcon from '@mui/icons-material/PlayArrow';
 
-import StreamedData from 'util/StreamedData/StreamedData';
 import DataPane from './DataPane';
 import TextSegmentController from './TextSegmentController';
 import { StatusMsg, StatusMsgSeverity } from './StatusMsg';
 // eslint-disable-next-line import/no-cycle
 import { Settings } from './Settings/Settings';
 import Terminal from './Terminal/Terminal';
+
+// console.log(version);
 
 declare global {
   interface String {
@@ -87,7 +84,6 @@ function formatDate(date: Date) {
 }
 
 export class App {
-  SerialPortType: typeof SerialPort | typeof SerialPortMock;
 
   settings: Settings;
 
@@ -107,8 +103,6 @@ export class App {
     true
   );
 
-  serialPort: null | SerialPort | SerialPortMock = null;
-
   portState = PortState.CLOSED;
 
   dataPane1: DataPane;
@@ -116,16 +110,6 @@ export class App {
   dataPane2: DataPane;
 
   rxData = '';
-
-  input: StreamedData;
-
-  newLineParser: NewLineParser;
-
-  buffer1: StreamedData;
-
-  ansiECParser: AnsiECParser;
-
-  output: StreamedData;
 
   txSegments: TextSegmentController;
 
@@ -152,8 +136,17 @@ export class App {
   // available to determine height
   testing: boolean;
 
+  port: SerialPort | null;
+
+  serialPortInfo: SerialPortInfo | null;
+
+  keepReading: boolean = true;
+
+  reader: ReadableStreamDefaultReader<Uint8Array> | null;
+
+  closedPromise: Promise<void> | null;
+
   constructor(
-    SerialPortType: typeof SerialPort | typeof SerialPortMock,
     testing = false
   ) {
     this.testing = testing;
@@ -168,31 +161,23 @@ export class App {
     this.dataPane1 = new DataPane();
     this.dataPane2 = new DataPane();
 
-    this.input = new StreamedData();
-    this.ansiECParser = new AnsiECParser();
-    this.buffer1 = new StreamedData();
-    this.newLineParser = new NewLineParser('\n');
-    this.output = new StreamedData();
-
     this.txSegments = new TextSegmentController();
     this.rxSegments = new TextSegmentController();
     // A mix of both TX and RX data. Displayed when the "COMBINED_TX_RX_PANE"
     // view configuration is selected.
     this.txRxSegments = new TextSegmentController();
 
-    this.addStatusBarMsg('Started NinjaTerm.', StatusMsgSeverity.INFO);
-    makeAutoObservable(this); // Make sure this near the end
+    this.port = null;
+    this.serialPortInfo = null;
+    this.reader = null;
+    this.closedPromise = null;
 
-    // WARNING: Make sure this is after makeAutoObservable()!!!
-    this.SerialPortType = SerialPortType;
+    console.log('Started NinjaTerm.');
+    makeAutoObservable(this); // Make sure this near the end
   }
 
   setSettingsDialogOpen(trueFalse: boolean) {
     this.settingsDialogOpen = trueFalse;
-    // If opening the settings dialog, also scan for ports
-    if (trueFalse) {
-      this.scanForPorts();
-    }
   }
 
   setCloseSettingsDialogOnPortOpenOrClose(trueFalse: boolean) {
@@ -202,96 +187,126 @@ export class App {
   /**
    * Scans the computer for available serial ports, and updates availablePortInfos.
    */
-  scanForPorts() {
-    this.SerialPortType.list()
-      .then((ports) => {
-        this.settings.setAvailablePortInfos(ports);
-        // Set the selected port, this doesn't fire automatically if setting
-        // the ports via code
-        if (ports.length > 0) {
-          this.settings.setSelectedPortPath(ports[0].path);
-        }
-        this.addStatusBarMsg(
-          `Port scan complete. Found ${ports.length} ports.`,
-          StatusMsgSeverity.INFO,
-          true
-        );
-        return 0;
+  async scanForPorts() {
+    // Prompt user to select any serial port.
+    console.log('hfhfh');
+    if ("serial" in navigator) {
+      // The Web Serial API is supported.
+      const localPort = await navigator.serial.requestPort();
+      console.log('Got local port, now setting state...');
+      runInAction(() => {
+        console.log('Setting this.port and this.serialPortInfo...');
+        this.port = localPort;
+        this.serialPortInfo = this.port.getInfo();
       })
-      .catch((error) => {
-        console.log(error);
-        this.addStatusBarMsg(`${error}`, StatusMsgSeverity.ERROR, true);
-      });
+    } else {
+      console.log('Not supported!');
+    }
+
+
+    // this.SerialPortType.list()
+    //   .then((ports) => {
+    //     this.settings.setAvailablePortInfos(ports);
+    //     // Set the selected port, this doesn't fire automatically if setting
+    //     // the ports via code
+    //     if (ports.length > 0) {
+    //       this.settings.setSelectedPortPath(ports[0].path);
+    //     }
+    //     this.addStatusBarMsg(
+    //       `Port scan complete. Found ${ports.length} ports.`,
+    //       StatusMsgSeverity.INFO,
+    //       true
+    //     );
+    //     return 0;
+    //   })
+    //   .catch((error) => {
+    //     console.log(error);
+    //     this.addStatusBarMsg(`${error}`, StatusMsgSeverity.ERROR, true);
+    //   });
   }
 
   async openPort() {
-    this.addStatusBarMsg('Opening port...', StatusMsgSeverity.INFO, true);
-    this.serialPort = new this.SerialPortType({
-      path: this.settings.selectedPortPath,
-      baudRate: this.settings.selectedBaudRate,
-      dataBits: this.settings.selectedNumDataBits as 5 | 6 | 7 | 8,
-      parity: this.settings.selectedParity as
-        | 'none'
-        | 'even'
-        | 'odd'
-        | 'mark'
-        | 'space',
-      stopBits: this.settings.selectedStopBits,
-      autoOpen: false, // Prevent serial port from opening until we call open()
-    });
+    // this.addStatusBarMsg('Opening port...', StatusMsgSeverity.INFO, true);
+    // this.serialPort = new this.SerialPortType({
+    //   path: this.settings.selectedPortPath,
+    //   baudRate: this.settings.selectedBaudRate,
+    //   dataBits: this.settings.selectedNumDataBits as 5 | 6 | 7 | 8,
+    //   parity: this.settings.selectedParity as
+    //     | 'none'
+    //     | 'even'
+    //     | 'odd'
+    //     | 'mark'
+    //     | 'space',
+    //   stopBits: this.settings.selectedStopBits,
+    //   autoOpen: false, // Prevent serial port from opening until we call open()
+    // });
 
-    // The open event is always emitted
-    this.serialPort.on('open', () => {
-      // open logic
-      this.setPortState(PortState.OPENED);
-      this.addStatusBarMsg(
-        'Port opened successfully.',
-        StatusMsgSeverity.OK,
-        true
-      );
-      // This will automatically close the settings window if the user is currently in it,
-      // clicks "Open" and the port opens successfully.
-      if (this.closeSettingsDialogOnPortOpenOrClose) {
-        this.setSettingsDialogOpen(false);
-      }
-    });
+    // navigator.serial.addEventListener("connect", (event) => {
+    //   // TODO: Automatically open event.target or warn user a port is available.
+    //   console.log('connect event called.');
+    // });
 
-    if (this.serialPort.isOpen) {
-      console.log('WARNING: Serial port already open!!!');
+    await this.port?.open({baudRate: this.settings.selectedBaudRate})
+    console.log('Serial port opened.');
+    this.setPortState(PortState.OPENED);
+      // this.addStatusBarMsg(
+      //   'Port opened successfully.',
+      //   StatusMsgSeverity.OK,
+      //   true
+      // );
+    // This will automatically close the settings window if the user is currently in it,
+    // clicks "Open" and the port opens successfully.
+    if (this.closeSettingsDialogOnPortOpenOrClose) {
+      this.setSettingsDialogOpen(false);
     }
-    this.serialPort.open((error) => {
-      if (error) {
-        console.log(error);
-        // Error already says "Error" at the start
-        this.addStatusBarMsg(`${error}`, StatusMsgSeverity.ERROR, true);
-      }
-    });
 
-    // Switches the port into "flowing mode"
-    this.serialPort.on('data', (data) => {
-      this.rxTerminal.parseData(data);
-      this.txRxTerminal.parseData(data);
-    });
+    this.keepReading = true;
+    this.closedPromise = this.readUntilClosed();
   }
 
-  closePort() {
-    if (!this.serialPort?.isOpen) {
-      console.log('closePort() called but port was not open.');
-      return;
-    }
-    this.serialPort?.close(() => {
-      this.setPortState(PortState.CLOSED);
-      // This will automatically close the settings window if the user is currently in it,
-      // clicks "Close" and the port closes successfully.
-      if (this.closeSettingsDialogOnPortOpenOrClose) {
-        this.setSettingsDialogOpen(false);
+  async readUntilClosed() {
+    // this.txRxTerminal.parseData(Buffer.from('s'));
+    while (this.port?.readable && this.keepReading) {
+      this.reader = this.port.readable.getReader();
+      try {
+        while (true) {
+          const { value, done } = await this.reader.read();
+          if (done) {
+            // reader.cancel() has been called.
+            break;
+          }
+          // value is a Uint8Array.
+          this.parseRxData(value);
+        }
+      } catch (error) {
+        // Handle error...
+      } finally {
+        // Allow the serial port to be closed later.
+        this.reader.releaseLock();
       }
-      this.addStatusBarMsg(
-        'Port successfully closed.',
-        StatusMsgSeverity.OK,
-        true
-      );
-    });
+    }
+
+    await this.port?.close();
+  }
+
+  parseRxData(value: Uint8Array) {
+    this.txRxTerminal.parseData(value);
+  }
+
+  async closePort() {
+    this.keepReading = false;
+    // Force reader.read() to resolve immediately and subsequently
+    // call reader.releaseLock() in the loop example above.
+    this.reader?.cancel();
+
+    if (this.closedPromise === null) {
+      throw Error('jfjfjf')
+    }
+    await this.closedPromise;
+
+    this.setPortState(PortState.CLOSED);
+    this.reader = null;
+    this.closedPromise = null;
   }
 
   setPortState(newPortState: PortState) {
@@ -328,7 +343,7 @@ export class App {
     }
   };
 
-  handleKeyPress(event: KeyboardEvent) {
+  async handleKeyPress(event: KeyboardEvent) {
     console.log('handleKeyPress() called. event=', event, this);
     if (this.portState === PortState.OPENED) {
       // Serial port is open, let's send it to the serial
@@ -360,19 +375,15 @@ export class App {
         console.log('Unsupported char!');
       }
       console.log('Writing to serial port. bytesToWrite=', bytesToWrite);
-      this.serialPort?.write(bytesToWrite, (error) => {
-        if (error) {
-          this.addStatusBarMsg(
-            `Could not write data to serial port. data=${event.key}, error=${error}.`,
-            StatusMsgSeverity.ERROR
-          );
-        } else {
-          // Sending was successful, increment TX count and insert sent data
-          // into TX and TXRX segments for showing in pane(s)
-          this.txTerminal.parseData(Buffer.from(bytesToWrite));
-          this.txRxTerminal.parseData(Buffer.from(bytesToWrite));
-        }
-      });
+      const writer = this.port?.writable?.getWriter();
+
+      const data = new Uint8Array([104, 101, 108, 108, 111]); // hello
+      await writer?.write(data);
+
+      // Allow the serial port to be closed later.
+      writer?.releaseLock();
+      this.txTerminal.parseData(Uint8Array.from(bytesToWrite));
+      this.txRxTerminal.parseData(Uint8Array.from(bytesToWrite));
     }
   }